' < sabrown |X| D14_03_12_13_04_00 | Remote benedict > Work over clang warnings with more flags'

+++ source/psh/do-net.c
@@ -356,7 +356,7 @@ static char *stop_time(char *et, int nc, double ti)
 /* COUNT_SPECS - return the number of specs */
 
 static int count_specs(hfspec *specs, size_t nsp)
-   {int n;
+   {size_t n;
 
     for (n = 0; (n < nsp) && (specs[n].running != UNLAUNCHED); n++);
